src/trosyn-sync/pyproject.toml
[project]
name = "trosyn-sync"
version = "0.1.0"
description = ""
authors = [
    {name = "ssentongoi",email = "ssentongoivan1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncio-dgram (>=2.2.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
